requirements.txt
torch==1.10.2
torchvision==0.11.1
tensorboard==2.4.1
pandas==1.1.5
opencv-python==4.5.1.48
scikit-learn~=0.24.1
matplotlib==3.2.1
numpy~=1.18.0
einops~=0.4.0
Pillow~=8.1.0
python-dotenv~=0.19.2
tqdm~=4.51.0
timm==0.5.4